feat(将LazyRow替换为RecyclerView):
1.新增音乐详情页
2.将Glide替换为Coil

+++ repository/src/main/java/com/brookes6/repository/model/RecommendSongModel.kt
@@ -29,7 +29,7 @@ data class RecommendSongModel(
         val alg: String = "",
         val ar: List<Author?>? = null,
         val al: SongImage? = null,
-        val mark: Int = 0,
+        val mark: Long = 0,
     )
 
     @Serializable
